+++ odoo/local-src/sf_rma/__manifest__.py
@@ -32,6 +32,7 @@
         "data/pricelist.xml",
         "data/ir_sequence.xml",
         "data/ir_config_parameter.xml",
+        "data/stock_picking_type.xml",
         "data/sale_order_type.xml",
     ],
 }