+++ HISTORY.rst
@@ -33,6 +33,7 @@ latest (unreleased)
 * Set 'add' as default value on mrp repair line
 * Install modules account_cutoff_accrual_base and account_cutoff_accrual_picking
 * Add RMA Settings to define RMA repair service product
+* Add RMA picking type
 
 **Bugfixes**
 
@@ -42,6 +43,10 @@ latest (unreleased)
 * Refresh customers data, error with the parent_id
 * Refresh suppliers data, error with the country and state
 * Fix warehouse creation, sequences names and company
+* Allow to cancel RMA
+* Require lot number if RMA product is tracked
+* Add lot number on RMA incoming picking
+* Set RMA default values
 
 **Build**